BACKEND - Correction ROUTE PUT POST (suppression fichier sans changement de contenu de post)

+++ backend/controllers/post.js
@@ -86,20 +86,28 @@ exports.updatePost = (req, res) => {
                 
                             return res.status(404).json({ message: "Erreur de traitement" });
                         })
+                } else if (req.body.imageUrl && req.body.imageUrl == "" && post.imageUrl) {
+                    const imageToBeDeleted = post.imageUrl.split('/images')[1];
+                    fs.unlink(`images/${imageToBeDeleted}`, (err) => {
+                        if (err) {
+                            console.error(err);
+                        }
+                        })
+                    
+                    post.update({ ...req.body }, { where: { id: req.params.postId } })
+                        .then(() => {
+                            
+                            return res.status(200).json({ message: "Votre post a bien été mis à jour" });
+                        })
+                        .catch(error => {
+                            console.error(error);
+                
+                            return res.status(404).json({ message: "Erreur de traitement" });
+                        })
                 } else {
 
                     post.update({ ...req.body }, { where: { id: req.params.postId } })
                         .then(() => {
-
-                            if (post.imageUrl) {
-                                const imageToBeDeleted = post.imageUrl.split('/images')[1];
-
-                                fs.unlink(`images/${imageToBeDeleted}`, (err) => {
-                                    if (err) {
-                                        console.error(err);
-                                    }
-                                    })
-                            }
                             
                             return res.status(200).json({ message: "Votre post a bien été mis à jour" });
                         })
